pyproject.toml
[project]
name = "text-to-image-generator"
version = "0.1.0"
description = "AI-powered text-to-image generation using Hugging Face models"
authors = [
    {name = "sai venkat",email = "saisam346@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-community (>=0.3.29,<0.4.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "huggingface-hub (>=0.34.4,<0.35.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "streamlit (>=1.49.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
